Add full detection to char array methods
